refactor(web): extract HtmlViewer's analytics fire-helpers into the file-viewer slice
Cluster B of the FileViewer.tsx decomposition: exportProgressRef and the
fireShareExport/fireArtifactToolbarClick/fireDrawToolbarClick/
fireArtifactHeaderClick/firePresentPopoverClick/fireCommentPopoverClick
fire-helpers move into features/file-viewer/hooks/useArtifactAnalytics.hooks.ts,
wired into the orchestrator via useWiredArtifactAnalytics from the slice
barrel. Structural move only, no logic changes.
